+++ android/src/main/java/com/dooboolab/RNIap/RNIapModule.java
@@ -1,3 +1,4 @@
+
 package com.dooboolab.RNIap;
 
 import android.app.Activity;
@@ -66,8 +67,6 @@ public class RNIapModule extends ReactContextBaseJavaModule {
   private IInAppBillingService mService;
   private BillingClient mBillingClient;
 
-  private boolean clientReady = false;
-
   private ServiceConnection mServiceConn = new ServiceConnection() {
     @Override public void onServiceDisconnected(ComponentName name) {
       mService = null;
@@ -109,7 +108,7 @@ public class RNIapModule extends ReactContextBaseJavaModule {
   }
 
   private void ensureConnection (final Promise promise, final Runnable callback) {
-    if (clientReady) {
+    if (mBillingClient != null && mBillingClient.isReady()) {
       callback.run();
       return;
     }
@@ -119,21 +118,24 @@ public class RNIapModule extends ReactContextBaseJavaModule {
     intent.setPackage("com.android.vending");
 
     final BillingClientStateListener billingClientStateListener = new BillingClientStateListener() {
+      private boolean bSetupCallbackConsumed = false;
+
       @Override
       public void onBillingSetupFinished(@BillingClient.BillingResponse int responseCode) {
-        if (responseCode == BillingClient.BillingResponse.OK ) {
-          Log.d(TAG, "billing client ready");
-          callback.run();
-          clientReady = true;
-        } else {
-          rejectPromiseWithBillingError(promise, responseCode);
+        if (!bSetupCallbackConsumed) {
+          bSetupCallbackConsumed = true;
+          if (responseCode == BillingClient.BillingResponse.OK ) {
+            Log.d(TAG, "billing client ready");
+            callback.run();
+          } else {
+            rejectPromiseWithBillingError(promise, responseCode);
+          }
         }
       }
 
       @Override
       public void onBillingServiceDisconnected() {
         Log.d(TAG, "billing client disconnected");
-        clientReady = false;
       }
     };
 
@@ -141,7 +143,6 @@ public class RNIapModule extends ReactContextBaseJavaModule {
       reactContext.bindService(intent, mServiceConn, Context.BIND_AUTO_CREATE);
       mBillingClient = BillingClient.newBuilder(reactContext).setListener(purchasesUpdatedListener).build();
       mBillingClient.startConnection(billingClientStateListener);
-      clientReady = true;
     } catch (Exception e) {
       promise.reject(E_NOT_PREPARED, e.getMessage(), e);
     }
@@ -172,8 +173,6 @@ public class RNIapModule extends ReactContextBaseJavaModule {
         return;
       }
     }
-
-    mBillingClient = null;
     promise.resolve(true);
   }
 
@@ -459,7 +458,7 @@ public class RNIapModule extends ReactContextBaseJavaModule {
         promise.reject(E_NETWORK_ERROR, "The service is disconnected (check your internet connection.)");
         break;
       case BillingClient.BillingResponse.USER_CANCELED:
-        promise.reject(E_USER_CANCELLED, "Cancelled.");
+        promise.reject(E_USER_CANCELLED, "Payment is Cancelled.");
         break;
       case BillingClient.BillingResponse.SERVICE_UNAVAILABLE:
         promise.reject(E_SERVICE_ERROR, "The service is unreachable. This may be your internet connection, or the Play Store may be down.");
@@ -471,7 +470,7 @@ public class RNIapModule extends ReactContextBaseJavaModule {
         promise.reject(E_ITEM_UNAVAILABLE, "That item is unavailable.");
         break;
       case BillingClient.BillingResponse.DEVELOPER_ERROR:
-        promise.reject(E_DEVELOPER_ERROR, "Google is indicating that we have an error in our code. Sorry about that.");
+        promise.reject(E_DEVELOPER_ERROR, "Google is indicating that we have some issue connecting to payment.");
         break;
       case BillingClient.BillingResponse.ERROR:
         promise.reject(E_UNKNOWN, "An unknown or unexpected error has occured. Please try again later.");