+++ index.d.ts
@@ -1,3 +1,5 @@
+import { EmitterSubscription } from 'react-native';
+
 import * as Apple from './apple'
 
 interface Common {
@@ -19,23 +21,22 @@ export interface Subscription<ID extends string> extends Common {
 
   introductoryPrice?: string
   introductoryPricePaymentModeIOS?: string
-  introductoryPriceNumberOfPeriods?: number
-  introductoryPriceSubscriptionPeriod: object
+  introductoryPriceNumberOfPeriodsIOS?: string
+  introductoryPriceSubscriptionPeriodIOS?: string
 
   subscriptionPeriodNumberIOS?: string
-  subscriptionPeriodUnitIOS?: number
+  subscriptionPeriodUnitIOS?: string
 
-  freeTrialPeriodAndroid?: string
-  introductoryPriceCyclesAndroid?: number
+  introductoryPriceCyclesAndroid?: string
   introductoryPricePeriodAndroid?: string
   subscriptionPeriodAndroid?: string
-  freeTrialPeriodAndroid: string
+  freeTrialPeriodAndroid?: string
 }
 
 export interface ProductPurchase {
   productId: string;
   transactionId: string;
-  transactionDate: string;
+  transactionDate: number;
   transactionReceipt: string;
   signatureAndroid?: string;
   dataAndroid?: string;
@@ -127,6 +128,17 @@ export function buySubscription(sku: string, oldSku?: string, prorationMode?: nu
 export function buyProduct(sku: string) : Promise<ProductPurchase>;
 
 /**
+ * Buy a product with offer
+ *
+ * @param {string} sku The product unique identifier
+ * @param {string} forUser An user identifier on your service (username or user id)
+ * @param {Apple.PaymentDiscount} withOffer The offer information
+ *
+ * @returns {Promise<void>}
+ */
+export function buyProductWithOfferIOS(sku: string, forUser: string, withOffer: Apple.PaymentDiscount) : Promise<void>;
+
+/**
  * Buy a product with a specified quantity (iOS only)
  * @param {string} sku The product's sku/ID
  * @param {number} quantity The amount of product to buy
@@ -190,4 +202,4 @@ export function validateReceiptAndroid(packageName: string, productId: string, p
   * Add IAP purchase event in ios.
  * @returns {callback(e: Event)}
  */
-export function addAdditionalSuccessPurchaseListenerIOS(fn: Function);
+export function addAdditionalSuccessPurchaseListenerIOS(fn: Function) : EmitterSubscription;